backend/requirements.txt
Flask
python-dotenv
gunicorn
moondream 
CORS
flask_cors
google-generativeai